Initial Upload of Code So far
Finished the main window and subpanels made of subpanels made of subpanels for the sudoku board. These are made of user controls but need to send events up to main window for the WFC algorithm to be called on by the main window.
   Based on what is allowed by neighboring cells, the sudokuInnerSquare enables or disabled buttons that are valid or invalid for the cell when double clicked on said cell.
Finished the Cell masking system and rules for dividing the cells as such.

For the Masking there are 2 sets of rules, single cell rules, and 2D rules.
All rules have an interface they stem from, the 2D and single cell rule's interfaces are different.
Single Cell Rules have a rule filter, rule dictionary, and rule set.
-Interface contains shared functions for getting valid values for a single cell and for testing if that cell contains a valid value
-Filter is a double array list that tells you if a value can exist for a single cell based on that filter. It contains a soft "suggestive" white list and a hard "preventative" blacklist. You can use operators to add the whitelists or subtract (add the blacklist)
-Set is a combination of multiple different single cell rules. It just combines all rules it has into one.
-Dictionary has a dictionary or one value to one filter. This is for if you want to use groupings for readability or for filtering values by existing values.
      example: a cell has value of 1, the dictionary contains 1 -> filter (whitelist [2,3] blacklist [4,5]). Dictionary would say 2 or 3 is a valid value to overwrite this cell, but not 4 or 5

2D rules
2D rules stem from the interface IRule_2D_Base
- Rule mask is built on an array of rule filters. It functions by giving what can be nearby a cell within it's range, if that rule mask is valid for that location.
- Rule Set and Rule Dictionary work similarly to the single cell counterpart, but with 2D locational values
- Singular Rule is a conversion from a 2D location to a single cell rule. This allows the 2D rules to use single cell rules.
- Axis Divider is a rule that creates a sub 2D array from the given array based upon the axis you want to divide, and a number or how many cells to divide for each group. It then performs rule calculations based on it's given 2D rules.

Possible Values Map is a 2D array of rule_Filters that contains functions for easier usage of combining rule_Filters and comparison of what is allowed at what location.

To Do -
Pass events from the inner square panel (0-9 button display) up to the main Form1 class in order to call updates for the 2D Rule Set for sudoku rules.
Create Wave Function Collapse algorithm that would help decide which cell is the least chaotic to the current board
Create return point panel on left side for returning board to a previous state should a chosen value not work for the panel
Change Rule_Filters to work on bit comparison rather than array's of allowed or disallowed values for quicker comparison. Would have issue converting bit comparison to other issues.