(Functional) Release
Board now shows inner square possibilities from the start.
For now, (Will be changed) Left Pane shows the squares you should collapse, e.g. the square with least possibilities. (This panel will be replaced in the future, but this feature will be implemented as a square highlighting function)
Now when you collapse, a square, the entire board will update
Now when you reset a square, the entire board will update