Insanity
Going in circles but general idea is this:

Remove any array cloning, this is taking a looooooot of time and processing power. Have the rules edit the tables directly.
This comes with many ramifications, specifically with the way rules returned tables.
Now instead, rules need to have a function called to load table data. Only the top rule needs to have the data loaded, other rules will trickle it down to subrules.
Possible values map has a subsection class that needs to inherit from possible values map. This will be just to interface a border or boundaries so we don't have to recreate the array.
Rule_2D_Base is now an abstract class instead of an interface. You can now clone Rules (But not 2D_Rules) and Possible Value Maps.

Future:
Possible values map now should have int[,] built into it. Rename it to values map.
Rename your variables.
Change the applyPossibleDataAboutPoint away from calling from a point p. This is redundant, especially when you convert p to its x and y only to then convert back to a point.
Also COMMENTS. :)